% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_summary.R
\name{chain_summary}
\alias{chain_summary}
\title{Summarize a single parameter from NIMBLE output}
\usage{
chain_summary(i, samples, chains = 1:3, cutoff = 0)
}
\arguments{
\item{i}{(numeric) index of parameter to summarize}

\item{samples}{(list) MCMC output from NIMBLE}

\item{chains}{(numeric vector) sequence of the number of chains; Default is 1:3}

\item{cutoff}{(numeric) where to cutoff chains in addition to the burnin; Default is 0}
}
\value{
A data.frame summarizing the mean, low, high, low tail, high tail, uncertainty, Rhat, and effective sample size (ESS) of a single parameter across all chains from nimble MCMC output.
}
\description{
Summarize a single parameter from NIMBLE output
}
