% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_covar.R
\name{cor_covar}
\alias{cor_covar}
\title{Plot correlations between covariates}
\usage{
cor_covar(
  covar,
  cov.names,
  cov.labels = cov.names,
  covs.int.factor = NA,
  out.path = "",
  out.name = "covariate-cor",
  color.threshold = 0.25
)
}
\arguments{
\item{covar}{(data.frame) data.frame holding covariate values and conus.grid.id}

\item{cov.names}{(character vector) column names from covar to plot}

\item{cov.labels}{(character vector) labels for covariate names, defaults to use column names as labels}

\item{covs.int.factor}{(character vector) name of columns in covar that has an interaction}

\item{out.path}{(character) path to save figure}

\item{out.name}{(character) file name to save figure}

\item{color.threshold}{(numeric) threshold above which to add color to plot}
}
\description{
Plot correlations between covariates
}
