% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_filter.R
\name{count_filter}
\alias{count_filter}
\title{Set up count data}
\usage{
count_filter(
  data,
  covs.mean,
  covs.sum,
  offset.area,
  age.use = c("adult", "metamorph", "juvenile", "larva", "egg mass", "NR", NA)
)
}
\arguments{
\item{data}{(data.frame) dataframe from species.data$obs containing observations from one dataset}

\item{covs.mean}{(character vector) vector of column names to use as detection covariates that should be averaged across passes (e.g., water temperature)}

\item{covs.sum}{(character vector) vector of column names to use as detection covariates that should be summed across passes (e.g., survey duration)}

\item{offset.area}{(character) name of column to use for area offset}

\item{age.use}{(character vector) vector of age categories to use; defaults to c("adult", "metamorph", "juvenile", "larva", "egg mass", "NR", NA)}
}
\value{
A dataframe containing clean and summarized observation data from one dataset.
}
\description{
Sets up count data for iSDM -- summarizes covariates. This function is wrapped within sppdata_for_nimble().
}
