% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_nimble.R
\name{data_for_nimble}
\alias{data_for_nimble}
\title{Format species data and covariate data for nimble}
\usage{
data_for_nimble(
  sp.data,
  covar,
  covs.z,
  sp.auto = T,
  coarse.grid = T,
  region,
  gridkey,
  spatRegion
)
}
\arguments{
\item{sp.data}{(list) output from sppdata_for_nimble()}

\item{covar}{(data.frame) dataframe containing ecological process covariates; must be scaled and contain columns for squared covariates with the column name "covariatename2"; conus.grid.id must match values from region$sp.grid.}

\item{covs.z}{(character vector) vector of column names from covar to use as covariates in the ecological process model; should contain names of both linear and quadratic covariates}

\item{sp.auto}{(logical) whether to use spatial autocorrelation in the model (T) or not (F); defaults to T}

\item{coarse.grid}{(logical) whether to use a coarse grid (T) or not (F); defaults to T}

\item{region}{(list) output from make_region()}

\item{gridkey}{(data.frame) contains conus.grid.ids and associated grid.ids to use for indexing in nimble}

\item{spatRegion}{(list) output from make_spatKey(); only required if coarse.grid == T}
}
\value{
(list) list of data and constants fully formatted for nimble
}
\description{
Format species data and covariate data for nimble
}
