% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_grid.R
\name{get_state_grid}
\alias{get_state_grid}
\title{Get state intersections with grid}
\usage{
get_state_grid(region, states)
}
\arguments{
\item{region}{(list) output from make_region()}

\item{states}{(sf) state polygons}
}
\value{
(data.frame) list of conus.grid.ids and which states they intersect with; can have multiple rows for one conus.grid.id if it overlaps multiple states
}
\description{
Get state intersections with grid
}
