% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_dup_records.R
\name{id_dup_records}
\alias{id_dup_records}
\title{Find duplicate records}
\usage{
id_dup_records(species.data)
}
\arguments{
\item{species.data}{(list) output from load_species_data()}
}
\value{
(vector) unique IDs of records that may be duplicates
}
\description{
Find PO records that might be duplicated in iNaturalist and in another dataset
}
