% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_covar.R
\name{load_covar}
\alias{load_covar}
\title{Load covariates}
\usage{
load_covar(sp.code, region)
}
\arguments{
\item{sp.code}{(character) 4-letter species code}

\item{region}{(list) output from make_region()}
}
\value{
(data.frame) dataframe holding covariate values for species grid
}
\description{
Load covariates
}
