% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species_data.R
\name{load_species_data}
\alias{load_species_data}
\title{Loads species data}
\usage{
load_species_data(
  sp.code,
  sp.code.all = sp.code,
  file.info,
  file.path,
  region,
  filter.region,
  year.start,
  year.end = as.numeric(format(Sys.Date(), "\%Y")),
  coordunc = 1000,
  coordunc_na.rm = T,
  spat.thin = F,
  statelines.rm = T,
  keep.conus.grid.id = region$sp.grid$conus.grid.id
)
}
\arguments{
\item{sp.code}{(character) 4-letter species code}

\item{sp.code.all}{(character) string of all species codes to pull for use in analysis, separated by |}

\item{file.info}{(data.frame) columns 'file.name', 'file.label', 'covar.mean', 'covar.sum', 'data.type', and 'PO.extent" describing the datasets to be read in for this species. See details for more information.}

\item{file.path}{(character) path where data files are located}

\item{region}{(sf list) region (output from make_region())}

\item{filter.region}{(logical) Only keep data within region (T) or not (F)}

\item{year.start}{(numeric) keep data during or after this year}

\item{year.end}{(numeric) keep data during or before this year; defaults to current year}

\item{coordunc}{(numeric) if data has coord.unc column (e.g., GBIF data), remove records with coord.unc > coordunc}

\item{coordunc_na.rm}{(logical) if data has coord.unc column (e.g., GBIF data), remove records with coord.unc == NA (T) or not (F)}

\item{spat.thin}{(logical) if data is PO, do spatial thinning at 2km x 2km grid cell (T) or not (F)}

\item{statelines.rm}{(logical) should state-specific PO data in cells that cross state lines be removed (T) or not (F)}

\item{keep.conus.grid.id}{(character vector) Which grid cells to keep (e.g., for cross validation); defaults to all grid cells in the region}
}
\value{
A list with two objects. The first object (locs) is a dataframe containing the coordinate locations of all observations. The second object (obs) is a list containing a dataframe for each dataset and the conus.grid.id location of each observation.
}
\description{
Loads species data and does initial cleaning necessary for iSDM
}
\details{
The columns in \code{file.info} must contain:
\itemize{
\item \code{file.name}: the name of the csv file to be read in
\item \code{file.label}: the label that should be used for the dataset. If multiple files have the same label, they will be appended together. Only files with the label "iNaturalist" will be treated as iNaturalist data.
\item \code{covar.mean}: detection covariates that should be averaged across passes; must match column name in data file
\item \code{covar.sum}: detection covariate(s) that should be summed across passes; must match column name in data file
\item \code{data.type}: data type of dataset; must be "PO", "DND", "or "Count"
\item \code{PO.extent}: describes the spatial extent of PO datasets; must be "CONUS" or a two-letter state abbreviation; NA for non-PO datasets
}
}
