% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CV_blocks.R
\name{make_CV_blocks}
\alias{make_CV_blocks}
\title{Make spatial blocks for cross validation}
\usage{
make_CV_blocks(region, rows, cols, k)
}
\arguments{
\item{region}{(list) output from make_region()}

\item{rows}{(integer) number of rows in blocks}

\item{cols}{(integer) number of cols in blocks}

\item{k}{(integer) number of folds to group blocks into}
}
\value{
sf object containing blocks
}
\description{
Make spatial blocks for cross validation
}
