% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_region.R
\name{make_region}
\alias{make_region}
\title{Make species region}
\usage{
make_region(
  rangelist,
  buffer,
  boundary = NULL,
  grid,
  sub = F,
  lat.lo = NA,
  lat.hi = NA,
  lon.lo = NA,
  lon.hi = NA,
  continuous = F,
  rm.clumps = F,
  clump.size = 20
)
}
\arguments{
\item{rangelist}{(list) list of ranges generated from get_range()}

\item{buffer}{(numeric) buffer size in meters}

\item{boundary}{(sf object) A polygon that says where to cut off the region (e.g., political borders)}

\item{grid}{(sf object) user-generated square or hexagonal grid with crs = 3857 and equal cell areas. Note that a coarse grid cannot be generated from square grid cells.}

\item{sub}{(logical) whether to buffer around centroid (sub = T) or around full range (sub = F)}

\item{lat.lo}{(numeric) lower boundary of latitude (degrees)}

\item{lat.hi}{(numeric) upper boundary of latitude (degrees)}

\item{lon.lo}{(numeric) lower boundary of longitude (degrees)}

\item{lon.hi}{(numeric) upper boundary of longitude (degrees)}

\item{continuous}{(logical) whether to keep only a continuous section of the range (T) or keep the whole range (F). See details for more information.}

\item{rm.clumps}{(logical) whether to remove clumps of cells of size clump.size. See details for more information.}

\item{clump.size}{(numeric) number of cells in clumps to keep}
}
\value{
A list containing ranges, the full region, sp.grid, and the boundary
}
\description{
Define the region for inference using a buffer around external range boundaries. Note that the function uses a Psuedo-mercator projection (EPSG:3857), which distorts real earth area at extreme latitudes. Further, the area of the cells are in Mercator meters, which may not reflect real earth meters in all areas. Users should determine whether this projection is suitable for their analysis.
}
\details{
The \code{continuous} and \code{rm.clups} arguments allow the user to specify whether they want a fully continuous region or a region without small groups of isolated cells. A continuous region, or at least a region without small clumps, can improve the performance of the spatial ICAR model. However, for species with a disjoint range, these arguments should be considered carefully as they may accidentally remove important parts of the species range.
}
