% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatkey.R
\name{make_spatkey}
\alias{make_spatkey}
\title{Make coarse spatial grid}
\usage{
make_spatkey(grid)
}
\arguments{
\item{grid}{(sf object) A grid in which one of the attributes is the conus.grid.id}
}
\value{
A list which contains:
\enumerate{
\item spatkey - a dataframe which contains a column for the conus.grid.id (fine grid) and a column for the matching spat.grid.id (coarse grid)
\item spat.grid - an sf dataframe which contains the spat.grid.id and the coarse grid geometry
}
}
\description{
Group a fine hexbin grid into groups of 7 hexbins to form a coarse spatial grid
}
\examples{
\dontrun{

spatial.grid <- make_spatkey(grid)

}
}
