% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species_data.R
\name{map_species_data}
\alias{map_species_data}
\title{Title}
\usage{
map_species_data(
  title,
  region,
  plot.range = T,
  plot.region = F,
  plot.cells = F,
  species.data,
  year.start,
  year.end = as.numeric(format(Sys.Date(), "\%Y")),
  plot = "samples",
  details = F,
  plot.blocks = F,
  blocks,
  out,
  plot.current = T,
  plot.uncertainty = F,
  plot.log = F,
  plot.exp = F,
  plot.change = F,
  proj = 0,
  proj.names = "",
  threshold = 0.5,
  coarse.grid,
  spatRegion
)
}
\arguments{
\item{spatRegion}{}
}
\description{
Title
}
