% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_code.R
\name{nimble_code}
\alias{nimble_code}
\title{Write code for nimble}
\usage{
nimble_code(
  data,
  constants,
  sp.auto,
  coarse.grid,
  path,
  Bprior = "dnorm(0,1)",
  block.out,
  zero_mean = T,
  tau = tau,
  rm.state = F
)
}
\arguments{
\item{data}{(list) data formatted for nimble}

\item{constants}{(list) constants formatted for nimble}

\item{sp.auto}{(logical) whether model should have spatial model (T) or not (F)}

\item{coarse.grid}{(logical) whether model uses the coarse spatial grid (T) or not (F)}

\item{path}{(character) where to save the nimble code}

\item{Bprior}{(character) distribution for beta priors; defaults to "dnorm(0,1)"}

\item{block.out}{(character or numeric) which block is excluded}

\item{zero_mean}{(logical) whether spatial parameter should include zero_mean restriction (T) or not (F); defaults to T}

\item{tau}{(character) the value of tau (precision) for the ICAR model}

\item{rm.state}{(logical) whether to remove state-specific effort (T) or not (F); defaults to F}
}
\value{
nimble code object
}
\description{
Write code for nimble based on contents of data and constants
}
