% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_inits.R
\name{nimble_inits}
\alias{nimble_inits}
\title{Sets up initial values for nimble}
\usage{
nimble_inits(data, constants, sp.auto, seed = as.numeric(Sys.time()))
}
\arguments{
\item{data}{(list) data formatted for nimble}

\item{constants}{(list) constants formatted for nimble}

\item{sp.auto}{(logical) whether model should have spatial model (T) or not (F)}

\item{seed}{(numeric) set seed to use for random number generation}
}
\value{
Initial values to be used in nimble code
}
\description{
Creates initial values for nimble based on content of 'data' and 'constants'
}
\examples{
\dontrun{

inits <- function(x){nimble_inits(data,
                                 constants,
                                 sp.auto = T,
                                 seed = x)}

}
}
