% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleParallel.R
\name{nimbleParallel}
\alias{nimbleParallel}
\title{Title}
\usage{
nimbleParallel(
  HPC,
  code,
  constants,
  data,
  inits,
  param,
  iter,
  burnin,
  thin,
  number,
  sp.code
)
}
\arguments{
\item{sp.code}{}
}
\description{
Title
}
