% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc.R
\name{plot_auc}
\alias{plot_auc}
\title{Plot AUC}
\usage{
plot_auc(auc, lines)
}
\arguments{
\item{auc}{(data.frame) dataframe containing auc information}

\item{lines}{(logical) whether to plot lines connecting in- and out-of-sample AUC (T) or not (F)}
}
\value{
ggplot of AUC
}
\description{
Plot AUC
}
