% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chains.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot chains}
\usage{
plot_chains(
  samples,
  data,
  cov.labs,
  cutoff = 0,
  plot = "B",
  chaincols = c(`1` = "hotpink1", `2` = "olivedrab3", `3` = "deepskyblue3")
)
}
\arguments{
\item{samples}{(list) output from nimble}

\item{data}{(list) list containing data that was input into nimble}

\item{cov.labs}{(data.frame) dataframe containing covariate column names ("covariate") and labels ("Label")}

\item{cutoff}{(numeric) where to cut off chains for plotting; defaults to 0}

\item{plot}{(character) which parameter to plot; defaults to "B"; options are "B" and "alpha"}

\item{chaincols}{(vector) colors to use for chains; defaults to pink, green, and blue}
}
\value{
ggplot object with plotted chains
}
\description{
Plot chains
}
