% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot chain convergence}
\usage{
plot_convergence(out)
}
\arguments{
\item{out}{(list) output from summarize_chains()}
}
\value{
ggplot of convergence metrics
}
\description{
Plot chain convergence
}
