% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{plot_covar}
\alias{plot_covar}
\title{Plot covariates}
\usage{
plot_covar(covar, region, cov.labs, scaled)
}
\arguments{
\item{covar}{(data.frame) dataframe containing covariate values for each conus.grid.id; conus.grid.id must match values from region$sp.grid.}

\item{region}{(list) region output from make_region()}

\item{cov.labs}{(data.frame) Labels for each covariate}

\item{scaled}{(logical) Are the covariate values scaled?}
}
\value{
Saves a map of scaled covariate values to the desired path.
}
\description{
Makes maps of covariate values.
}
