% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pars.R
\name{plot_pars}
\alias{plot_pars}
\title{Plot parameter estimates}
\usage{
plot_pars(out, plot.group = "process", plot.type = "full", cov.labs, title)
}
\arguments{
\item{out}{(list) output from summarize_chains()}

\item{plot.group}{(character) which type of parameters to plot; options are "process", "observation", or "dataset"; defaults to "process"}

\item{plot.type}{(character) which type of plot to make; options are "full" or "cv"; defaults to "full}

\item{cov.labs}{(data.frame) dataframe containing covariate column names ("covariate") and labels ("Label")}

\item{title}{(character) title for plot}
}
\value{
A plot of parameter estimates
}
\description{
Plot parameter estimates
}
