% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_for_nimble.R
\name{PO_for_nimble}
\alias{PO_for_nimble}
\title{Format PO data for nimble}
\usage{
PO_for_nimble(POdata, covariates, rename = "")
}
\arguments{
\item{POdata}{(data.frame) dataframe containing PO data}

\item{covariates}{(data.frame) dataframe containing covariates for PO effort}

\item{rename}{(character) character suffix to add to data and constants object names; defaults to empty string}
}
\value{
A list containing data and constants for nimble code
}
\description{
Creates data and constants necessary to fit PO data in nimble code
}
