% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nimbleMCMC.R
\name{run_nimbleMCMC}
\alias{run_nimbleMCMC}
\title{run_nimbleMCMC}
\usage{
run_nimbleMCMC(info, code, constants, data, param, iter, burnin, thin)
}
\arguments{
\item{info}{(list) A list which contains 2 elements: info$inits which contains the initial values for a single MCMC chain, and info$seed which is a single numeric to set the seed for the chain.}

\item{code}{(nimbleCode)}

\item{constants}{(nimbleData)}

\item{data}{(nimbleData)}

\item{param}{(list?)}

\item{iter}{(numeric) The number of iterations to run the chain for}

\item{burnin}{(numeric) The number of iterations to discard as burnin}

\item{thin}{(numeric) The number of iterations to thin by}
}
\value{
An coda::mcmc object of the posterior samples
}
\description{
A self-contained NIMBLE workflow to configure, build, compile, and run a NIMBLE model for a single chain.
}
