% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_this.R
\name{scale_this}
\alias{scale_this}
\title{Scale values}
\usage{
scale_this(x)
}
\arguments{
\item{x}{(numeric vector) A vector of values to re-scale}
}
\value{
A vector of scaled values with mean 0 and standard deviation 1
}
\description{
A function to re-scale values to a distribution with mean 0 and standard deviation 1
}
\examples{
x <- rnorm(1000, mean = 10, sd = 5)

scale_this(x)

}
