% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_covar.R
\name{select_covar}
\alias{select_covar}
\title{Select covariates that have low correlations}
\usage{
select_covar(covs, threshold = 0.4)
}
\arguments{
\item{covs}{(data.frame) dataframe holding pool of covariates to choose from}

\item{threshold}{(numeric) threshold for correlations above which covariates should be removed; defaults to 0.4}
}
\value{
(character vector) vector of covariate names to remove
}
\description{
Selectively remove covariates from a pool of covariates until all high correlations are gone.
}
