% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppdata_for_nimble.R
\name{sppdata_for_nimble}
\alias{sppdata_for_nimble}
\title{Format species data for nimble}
\usage{
sppdata_for_nimble(
  species.data,
  region,
  file.info,
  covar,
  stategrid,
  statelines.rm = T,
  covs.inat = c("traveltime", "density", "n.inat"),
  covs.PO = c("traveltime", "density"),
  DND.maybe = 1,
  keep.conus.grid.id,
  min.visits.incl = 3
)
}
\arguments{
\item{species.data}{(list) output from load_species_data()}

\item{region}{(list) output from make_region()}

\item{file.info}{(data.frame) columns 'file.name', 'file.label', 'covar.mean', 'covar.sum', 'data.type', and 'PO.extent" describing the datasets to be read in for this species. See details for more information.}

\item{covar}{(data.frame) dataframe containing effort covariates for PO data for each conus.grid.id; conus.grid.id must match values from region$sp.grid}

\item{stategrid}{(data.frame) output from get_state_grid()}

\item{statelines.rm}{(logical) should state-specific PO data in cells that cross state lines be removed (T) or not (F)}

\item{covs.inat}{(character vector) vector of column names (from covar) to use for effort covariates for iNat data}

\item{covs.PO}{(character vector) vector of column names (from covar) to use for effort covariates for non-iNat PO data}

\item{DND.maybe}{(numeric) whether to treat maybe detections as detections (1) or non-detections(0); defaults to 1}

\item{keep.conus.grid.id}{(vector) vector of conus.grid.ids to use to fit the model}

\item{min.visits.incl}{(numeric) minimum number (inclusive) of median visits per site to use n-mixture or occupancy models}
}
\value{
(list) list of species data partially formatted for nimble, needs to be input into data_for_nimble before use in nimble
}
\description{
Format species data for nimble
}
\details{
The columns in \code{file.info} must contain:
\itemize{
\item \code{file.name}: the name of the csv file to be read in
\item \code{file.label}: the label that should be used for the dataset. If multiple files have the same label, they will be appended together. Only files with the label "iNaturalist" will be treated as iNaturalist data.
\item \code{covar.mean}: detection covariates that should be averaged across passes; must match column name in data file
\item \code{covar.sum}: detection covariate(s) that should be summed across passes; must match column name in data file
\item \code{data.type}: data type of dataset; must be "PO", "DND", "or "Count"
\item \code{PO.extent}: describes the spatial extent of PO datasets; must be "CONUS" or a two-letter state abbreviation; NA for non-PO datasets
}

Note that survey detection covariates are scaled within this function; PO effort covariates must be scaled prior to this function.
}
