% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_for_nimble.R
\name{z_for_nimble}
\alias{z_for_nimble}
\title{Set up process data for nimble}
\usage{
z_for_nimble(covariates)
}
\arguments{
\item{covariates}{(data.frame) dataframe containing process covariates}
}
\value{
A list containing data and constants for nimble code
}
\description{
Creates data and constants necessary to fit process covariates in nimble code
}
