% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_state_ind.R
\name{add_state_ind}
\alias{add_state_ind}
\title{Add state indicator variables}
\usage{
add_state_ind(
  species.data,
  region,
  gridkey,
  constants,
  covs.inat = NA,
  obsc.state = NA,
  keep.conus.grid.id = gridkey$conus.grid.id[which(gridkey$group == "train")]
)
}
\arguments{
\item{species.data}{(list) output from load_species_data()}

\item{region}{(list) output from make_region()}

\item{gridkey}{(data.frame) links conus.grid.id and grid.id}

\item{constants}{(list) output from data_for_nimble()}

\item{covs.inat}{(character vector) vector of column names (from covar) to use for effort covariates for iNat data}

\item{obsc.state}{(character vector) abbreviations of states where species has taxon geoprivacy}

\item{keep.conus.grid.id}{(character vector) vector of conus.grid.ids to keep; default comes from gridkey}
}
\value{
(list) list of species data partially formatted for nimble, needs to be input into data_for_nimble before use in nimble
}
\description{
Add state indicator variables
}
