% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_covar.R
\name{cor_covar}
\alias{cor_covar}
\title{Plot correlations between covariates}
\usage{
cor_covar(
  covar,
  cov.names,
  cov.labels = cov.names,
  cov.levels = cov.labels,
  out.path = "",
  out.name = "covariate-cor",
  color.threshold = 0.25
)
}
\arguments{
\item{covar}{(data.frame) data.frame holding covariate values and conus.grid.id}

\item{cov.names}{(character vector) column names from covar to plot}

\item{cov.labels}{(character vector) labels for covariate names, defaults to use column names as labels}

\item{cov.levels}{(character vector) column names in order they should appear on plot}

\item{out.path}{(character) path to save figure}

\item{out.name}{(character) file name to save figure}

\item{color.threshold}{(numeric) threshold above which to add color to plot; default is 0.25}
}
\value{
Saves a plot showing correlations between covariates
}
\description{
Calculates and plots correlations between covariates.
}
\examples{
\dontrun{
cor_covar(covar,
          cov.names = c("cwd", "permwater"),
          cov.labels = c("CWD", "Permanent water"),
          out.path = "outputs/",
          out.name = "covs-cor")
}
}
