% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_covar.R
\name{cor_covar}
\alias{cor_covar}
\title{Plot correlations between covariates}
\usage{
cor_covar(covar, cov.labs, cov.levels = cov.labels, color.threshold = 0.25)
}
\arguments{
\item{covar}{(data.frame) data.frame holding covariate values and conus.grid.id}

\item{cov.labs}{(data.frame) Labels for each covariate}

\item{cov.levels}{(character vector) column names in order they should appear on plot}

\item{color.threshold}{(numeric) threshold above which to add color to plot; default is 0.25}
}
\value{
Saves a plot showing correlations between covariates
}
\description{
Calculates and plots correlations between covariates.
}
