% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_nimble.R
\name{data_for_nimble}
\alias{data_for_nimble}
\title{Format species data and covariate data for nimble}
\usage{
data_for_nimble(
  sp.data,
  covar,
  covs.z,
  sp.auto = T,
  coarse.grid = T,
  region,
  area = F,
  process.intercept = T,
  gridkey,
  spatRegion
)
}
\arguments{
\item{spatRegion}{}
}
\description{
Format species data and covariate data for nimble
}
