% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpus_per_task.R
\name{get_cpus_per_task}
\alias{get_cpus_per_task}
\title{Get CPUs per task (SLURM)}
\usage{
get_cpus_per_task()
}
\value{
The number of CPUs allocated to the SLURM task
}
\description{
Get CPUs per task (SLURM)
}
