% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_derived.R
\name{get_derived}
\alias{get_derived}
\title{Extract and calculate derived quantities}
\usage{
get_derived(
  samples,
  data,
  project,
  coarse.grid,
  spatRegion,
  pathToProj = NULL,
  sp.auto = T
)
}
\arguments{
\item{samples}{(list of mcmc objects) output from fitting a NIMBLE model}

\item{data}{(NIMBLE data) data created and formatted for NIMBLE}

\item{project}{(numeric) the number of projections}

\item{coarse.grid}{(logical) whether or not a coarse grid was used to fit the spatial model}

\item{spatRegion}{(list) object created from make_spatkey() and used to fit the spatial model}

\item{pathToProj}{(character) path to R script where projections are set up}

\item{sp.auto}{(logical) whether or not spatial autocorrelation was used}
}
\value{
A dataframe of the original MCMC samples with estimated psi and lambda, psi, and XB for each projection (if projections are specified)
}
\description{
Extract and calculate derived quantities
}
