% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get species ranges for reference}
\usage{
get_range(range.path, range.name = paste0("range", 1:length(range.path)), crs)
}
\arguments{
\item{range.path}{A vector of paths leading to species ranges}

\item{range.name}{A vector of names for each range}

\item{crs}{EPSG code for desired crs of output}
}
\value{
A list containing ranges as sf objects
}
\description{
Get species ranges for reference
}
\examples{
\dontrun{
range <- get_range('ANMI')
}
}
