% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get GAP and IUCN species ranges for reference}
\usage{
get_range(sp.code)
}
\arguments{
\item{sp.code}{}
}
\value{
A list containing the GAP and IUCN species ranges (if pre-downloaded and available)
}
\description{
Get GAP and IUCN species ranges for reference
}
\examples{
\dontrun{
range <- get_range('ANMI')
}
}
