% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get species ranges for reference}
\usage{
get_range(range.path, range.name = paste0("range", 1:length(range.path)), crs)
}
\arguments{
\item{range.path}{(character vector) A vector of paths leading to species ranges}

\item{range.name}{(character vector) A vector of names for each range}

\item{crs}{(numeric) EPSG code for desired coordinate reference system (CRS) of output}
}
\value{
A list containing ranges as sf objects
}
\description{
Assembles pre-downloaded species range shapefiles (e.g., from IUCN or GAP)
}
\examples{
\dontrun{
# Predownload range boundaries of interest

# Provide the path to each boundary
range.path <- c("GAP/aAZTOx_CONUS_Range_2001v1/",
                "IUCN/")

# Give each boundary a name
range.name <- c("GAP", "IUCN")

rangelist <- get_range(range.path,
                       range.name,
                       crs = 4326)
}
}
