% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species_data.R
\name{load_species_data}
\alias{load_species_data}
\title{Load species data}
\usage{
load_species_data(
  sp.code,
  file.name,
  file.label,
  file.path,
  keep.subsp,
  keep.cols,
  region,
  filter.region,
  year.start,
  year.end,
  coordunc = 1000,
  coordunc_na.rm = T,
  spat.thin = F,
  keep.conus.grid.id = region$sp.grid$conus.grid.id
)
}
\arguments{
\item{sp.code}{4-letter species code}

\item{file.name}{vector of data file names to read in}

\item{file.label}{vector of labels for data files}

\item{file.path}{path where data files are located}

\item{keep.subsp}{whether to keep records of subsp (T) or not (F)}

\item{keep.cols}{list of column names to keep for each data file}

\item{region}{region (output from make_region())}

\item{filter.region}{Only keep data within region (T) or not (F)}

\item{year.start}{keep data during or after this year}

\item{year.end}{keep data during or before this year}

\item{coordunc}{If data has coord.unc column (e.g., GBIF data), remove records with coord.unc > coordunc}

\item{coordunc_na.rm}{If data has coord.unc column (e.g., GBIF data), remove records with coord.unc == NA (T) or not (F)}

\item{spat.thin}{If data is PO, do spatial thinning at 2km x 2km grid cell}

\item{keep.conus.grid.id}{Which grid cells to keep (e.g., for cross validation); defaults to all grid cells in the region}
}
\description{
Load species data
}
\examples{
\dontrun{
range <- get_range('ANMI')
}
}
