% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species_data.R
\name{load_species_data}
\alias{load_species_data}
\title{Loads species data}
\usage{
load_species_data(
  sp.code,
  sp.code.all,
  file.info,
  file.path,
  region,
  filter.region,
  year.start,
  year.end = as.numeric(format(Sys.Date(), "\%Y")),
  coordunc = 1000,
  coordunc_na.rm = T,
  spat.thin = F,
  keep.conus.grid.id = region$sp.grid$conus.grid.id
)
}
\arguments{
\item{sp.code}{(character) 4-letter species code}

\item{sp.code.all}{(character) string of all species codes to pull for use in analysis, separated by |}

\item{file.info}{(data.frame) columns 'file.name', 'file.label', 'covar.mean', and 'covar.sum' describing the datasets to be read in for this species}

\item{file.path}{(character) path where data files are located}

\item{region}{(sf list) region (output from make_region())}

\item{filter.region}{(logical) Only keep data within region (T) or not (F)}

\item{year.start}{(numeric) keep data during or after this year}

\item{year.end}{(numeric) keep data during or before this year; defaults to current year}

\item{coordunc}{(numeric) if data has coord.unc column (e.g., GBIF data), remove records with coord.unc > coordunc}

\item{coordunc_na.rm}{(logical) if data has coord.unc column (e.g., GBIF data), remove records with coord.unc == NA (T) or not (F)}

\item{spat.thin}{(logical) if data is PO, do spatial thinning at 2km x 2km grid cell (T) or not (F)}

\item{keep.conus.grid.id}{(character vector) Which grid cells to keep (e.g., for cross validation); defaults to all grid cells in the region}
}
\value{
A list with two objects. The first object (locs) is a dataframe containing the coordinate locations of all observations. The second object (obs) is a list containing a dataframe for each dataset and the conus.grid.id location of each observation.
}
\description{
Loads species data and does initial cleaning necessary for iSDM
}
\examples{
\dontrun{

species.data <- load_species_data(sp.code = "ANMI",
                sp.code.all = "ANMI",
                file.name = c("name-of-file.csv"),
                file.label = c("File Label"),
                file.path = "DATA SWAMP/data-ready/",
                keep.cols = list(c("duration")),
                region = region,
                filter.region = T,
                year.start = 1994,
                year.end = 2004,
                coordunc = 1000,
                coordunc_na.rm = T,
                spat.thin = F,
                keep.conus.grid.id = region$sp.grid$conus.grid.id

}
}
