% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_region.R
\name{make_region}
\alias{make_region}
\title{Make species region}
\usage{
make_region(
  rangelist,
  buffer,
  boundary,
  grid,
  crs = 3857,
  sub = F,
  lat.lo = NA,
  lat.hi = NA,
  lon.lo = NA,
  lon.hi = NA,
  continuous = F,
  rm.clumps = F,
  clump.size = 20,
  cell.size = 24999998
)
}
\arguments{
\item{rangelist}{List of ranges}

\item{buffer}{Buffer size, unit depends on crs}

\item{boundary}{sf object that says where to cut off the region}

\item{grid}{sf object with grid}

\item{crs}{default is 3857 so that buffer size can be meters}

\item{sub}{whether to buffer around centroid (sub = T) or around full range (sub = F)}

\item{lat.lo}{lower boundary of latitude (degrees)}

\item{lat.hi}{upper boundary of latitude (degrees)}

\item{lon.lo}{upper boundary of longitude (degrees)}

\item{continuous}{whether to keep only a continuous section of the range (T) or keep the whole range (F)}

\item{rm.clumps}{whether to remove clumps of cells of size clump.size}

\item{clump.size}{number of cells in clumps to keep}
}
\value{
A list containing ranges, the full region, sp.grid, and the boundary
}
\description{
Make species region
}
\examples{
\dontrun{
range <- get_range('ANMI')
}
}
