% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species_data.R
\name{map_species_data}
\alias{map_species_data}
\title{Map species data}
\usage{
map_species_data(
  title,
  region,
  plot = "samples",
  plot.range = T,
  plot.region = F,
  plot.cells = F,
  species.data,
  year.start = 1900,
  year.end = as.numeric(format(Sys.Date(), "\%Y")),
  details = F,
  blocks = NULL,
  out,
  plot.current = T,
  plot.uncertainty = F,
  transform = "none",
  plot.change = F,
  proj.n = 0,
  proj.names = "",
  threshold = 0.5
)
}
\arguments{
\item{title}{(character) string to use as title of map}

\item{region}{(list) output from make_region()}

\item{plot}{(character) indicates what should be mapped; possible values are "samples", "lambda", "psi", "spat", "boundary", "XB"}

\item{plot.range}{(logical) whether to plot range outlines contained in region$range (T) or not (F); defaults to T}

\item{plot.region}{(logical) whether to plot region contained in region$region (T) or not (F); defaults to F}

\item{plot.cells}{(logical) whether to plot grid cells contained in region$sp.grid (T) or not (F); defaults to F}

\item{species.data}{(list) output from load_species_data(); only required if plot == "samples"}

\item{year.start}{(numeric) plot data from during or after this year; defaults to 1900; only required if plot == "samples"}

\item{year.end}{(numeric) plot data from during or before this year; defaults to current year; only required if plot == "samples"}

\item{details}{(logical) whether to add dataset details to map (T) or not (F); defaults to F; only required if plot == "samples"}

\item{blocks}{(sf) sf object containing blocks, e.g., output from cv_spatial(), out$blocks; only required if plot.blocks == T}

\item{out}{(list) summarized output from nimble model; output from summarize_samples(); only required if plot != "samples"}

\item{plot.current}{(logical) whether to plot the current output (T) or projections (F); defaults to T; only required if plot != "samples"}

\item{plot.uncertainty}{(logical) plot relative uncertainty ("unc.rel"), absolute uncertainty ("unc.range"), or mean values (F); defaults to F; only required if plot != "samples"}

\item{transform}{(character) transform ("log" or "exp") plotted values, or not ("none"); defaults to "none"}

\item{plot.change}{(logical) whether to plot the projected value (F) or the projected change from current (T); defaults to F; only required if plot.current == F}

\item{proj.n}{(numeric) number of projection scenarios to plot}

\item{proj.names}{(character vector) names for each projection; only required if plot.current == F}

\item{threshold}{(numeric) value between 0 and 1 to use as occupancy probability threshold for drawing range boundary; defaults to 0.5; only required if plot == "boundary"}

\item{plot.blocks}{(logical) whether to add blocks to map (T) or not (F); defaults to F; only required if plot == "samples"}
}
\value{
ggplot object containing map of desired data
}
\description{
Map species data
}
