% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_code.R
\name{nimble_code}
\alias{nimble_code}
\title{Write code for nimble}
\usage{
nimble_code(
  data,
  constants,
  sp.auto,
  coarse.grid,
  path,
  min.visits.incl = 3,
  Bpriordist = "dnorm",
  Bpriorvar1 = 0,
  Bpriorvar2 = 1,
  block.out,
  zero_mean = T,
  tau = tau,
  rm.state = F,
  occupancy = T
)
}
\arguments{
\item{data}{(list) data formatted for nimble}

\item{constants}{(list) constants formatted for nimble}

\item{sp.auto}{(logical) whether model should have spatial model (T) or not (F)}

\item{coarse.grid}{(logical) whether model uses the coarse spatial grid (T) or not (F)}

\item{path}{(character) where to save the nimble code}

\item{min.visits.incl}{(numeric) minimum number of median visits per site (inclusive) to use occupancy or N-mixture model; default is 3}

\item{Bpriordist}{(character) distribution for beta priors; defaults to "dnorm"}

\item{Bpriorvar1}{(numeric) first parameter for beta priors; defaults to 0}

\item{Bpriorvar2}{(numeric) second parameter for beta priors; defaults to 1}

\item{block.out}{(character or numeric) which block is excluded}

\item{zero_mean}{(logical) whether spatial parameter should include zero_mean restriction (T) or not (F); defaults to T}

\item{tau}{(character) the value of tau (precision) for the ICAR model}

\item{rm.state}{(logical) whether to remove state-specific effort (T) or not (F); defaults to F}

\item{occupancy}{(logical) whether to use occupancy models (T) or replace them with Bernoullis (F)}
}
\value{
nimble code object
}
\description{
Write code for nimble based on contents of data and constants
}
