% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleParallel.R
\name{nimbleParallel}
\alias{nimbleParallel}
\title{nimbleParallel}
\usage{
nimbleParallel(
  cores = 3,
  code,
  constants,
  data,
  inits,
  param,
  iter,
  burnin,
  thin
)
}
\arguments{
\item{cores}{(numeric) The number of cores (and, in turn, chains) to run in parallel. Default = 3.}

\item{code}{(nimbleCode)}

\item{constants}{(nimbleConstants)}

\item{data}{(nimbleData)}

\item{inits}{(function)}

\item{param}{(list?)}

\item{iter}{(numeric) The number of iterations to run in each chain}

\item{burnin}{(numeric) The number of iterations to discard as burnin}

\item{thin}{(numeric) The number of iterations to thin by}
}
\value{
A list of mcmc objects. Each element of the list is one chain of samples saved as a coda::mcmc object.
}
\description{
Run NIMBLE chains in parallel using a local cluster
}
