% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{plot_covar}
\alias{plot_covar}
\title{Plot covariates}
\usage{
plot_covar(
  covar,
  region,
  cov.names,
  cov.labels = cov.names,
  covs.int.factor = NA,
  out.path = "",
  out.name = "covariate-map"
)
}
\arguments{
\item{covar}{(data.frame) dataframe holding covariate values and conus.grid.id}

\item{region}{(list) region output from make_region()}

\item{cov.names}{(character vector) column names from covar to plot}

\item{cov.labels}{(character vector) labels for covariate names, defaults to use column names as labels}

\item{covs.int.factor}{(character) name of columns in covar that has an interaction}

\item{out.path}{(character) path to save figure}

\item{out.name}{(character) file name to save figure}
}
\value{
Saves a map of scaled covariate values to the desired path.
}
\description{
Makes maps of scaled covariate values.
}
\examples{
\dontrun{


plot_covar(covar,
           region,
           cov.names = c("cwd", "permwater"),
           cov.labels = c("CWD", "Permanent water"),
           out.path = "outputs/",
           out.name = "covs-map")

}
}
