% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{plot_covar}
\alias{plot_covar}
\title{Plot covariates}
\usage{
plot_covar(
  covar,
  region,
  cov.names,
  cov.labels = cov.names,
  covs.int.factor = NA,
  out.path = "",
  out.name = "covariate-map"
)
}
\arguments{
\item{covar}{data.frame holding covariate values and conus.grid.id}

\item{region}{region output from make_region()}

\item{cov.names}{column names from covar to plot}

\item{cov.labels}{labels for covariate names, defaults to use column names as labels}

\item{covs.int.factor}{name of columns in covar that has an interaction}

\item{out.path}{path to save figure}

\item{out.name}{file name to save figure}
}
\description{
Plot covariates
}
\examples{
\dontrun{
range <- get_range('ANMI')
}
}
