% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{plot_covar}
\alias{plot_covar}
\title{Plot covariates}
\usage{
plot_covar(covar, region, cov.labs, scaled)
}
\arguments{
\item{covar}{(data.frame) dataframe holding covariate values and conus.grid.id}

\item{region}{(list) region output from make_region()}

\item{cov.labs}{(data.frame) Labels for each covariate}

\item{scaled}{(logical) Are the covariate values scaled?}
}
\value{
Saves a map of scaled covariate values to the desired path.
}
\description{
Makes maps of scaled covariate values.
}
\examples{
\dontrun{
plot_covar(covar,
           region,
           cov.names = c("cwd", "permwater"),
           cov.labels = c("CWD", "Permanent water"),
           out.path = "outputs/",
           out.name = "covs-map")

}
}
