% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pars.R
\name{plot_pars}
\alias{plot_pars}
\title{Plot parameter estimates}
\usage{
plot_pars(out, plot.type = "full", cov.labs)
}
\arguments{
\item{out}{(data.frame) output from summarize_chains() containing correct parameters}

\item{plot.type}{(character) which type of plot to make; options are "full" or "cv"; defaults to "full}

\item{cov.labs}{(data.frame) dataframe containing covariate column names ("covariate") and labels ("Label")}
}
\value{
A plot of parameter estimates
}
\description{
Plot parameter estimates
}
