% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posteriors.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot posterior distributions}
\usage{
plot_posteriors(
  samples,
  data,
  cov.labs,
  cutoff = 0,
  plot = "B",
  Bpriordist = "dnorm",
  Bpriorvar1 = 0,
  Bpriorvar2 = 1,
  chaincols = c(`1` = "hotpink1", `2` = "olivedrab3", `3` = "deepskyblue3")
)
}
\arguments{
\item{samples}{(list) output from nimble}

\item{data}{(list) list containing data that was input into nimble}

\item{cov.labs}{(data.frame) dataframe containing covariate column names ("covariate") and labels ("Label")}

\item{cutoff}{(numeric) where to cut off chains for plotting; defaults to 0}

\item{plot}{(character) which parameter to plot; defaults to "B"; options are "B", "alpha", and "tau"}

\item{Bpriordist}{(character) distribution for beta priors; defaults to "dnorm"; options are "dnorm" or "dunif"}

\item{Bpriorvar1}{(numeric) first parameter for beta priors; defaults to 0}

\item{Bpriorvar2}{(numeric) second parameter for beta priors; defaults to 1}

\item{chaincols}{(vector) colors to use for chains; defaults to pink, green, and blue}
}
\value{
ggplot object with plotted posterior distributions
}
\description{
Plot posterior distributions
}
