% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_for_nimble.R
\name{PO_for_nimble}
\alias{PO_for_nimble}
\title{Format PO data for nimble}
\usage{
PO_for_nimble(POdata, covariates, rename = "")
}
\value{
A list containing ranges as sf objects
}
\description{
Format PO data for nimble
}
