% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppdata_for_nimble.R
\name{sppdata_for_nimble}
\alias{sppdata_for_nimble}
\title{Format species data for nimble}
\usage{
sppdata_for_nimble(
  species.data,
  region,
  data.type,
  PO.extent,
  covar,
  covs.inat = c("traveltime", "density", "n.inat"),
  covs.PO = c("traveltime", "density"),
  covs.mean,
  covs.sum,
  offset.area,
  DND.maybe = 1,
  keep.conus.grid.id,
  min.visits.incl = 3
)
}
\arguments{
\item{min.visits.incl}{}
}
\description{
Format species data for nimble
}
