% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppdata_for_nimble.R
\name{sppdata_for_nimble}
\alias{sppdata_for_nimble}
\title{Format species data for nimble}
\usage{
sppdata_for_nimble(
  species.data,
  region,
  data.type,
  PO.extent,
  covar,
  covs.inat = c("traveltime", "density", "n.inat"),
  covs.PO = c("traveltime", "density"),
  covs.mean,
  covs.sum,
  offset.area,
  DND.maybe = 1,
  keep.conus.grid.id,
  min.visits.incl = 3
)
}
\arguments{
\item{species.data}{(list) output from load_species_data()}

\item{region}{(list) output from make_region()}

\item{data.type}{(character vector) vector of data types for each dataset in species.data}

\item{PO.extent}{(character vector) vector of extents for each PO dataset (CONUS or state code); use NA for non-PO datasets}

\item{covar}{(data.frame) dataframe containing covariates for PO data}

\item{covs.inat}{(character vector) vector of column names (from covar) to use for effort covariates for iNat data}

\item{covs.PO}{(character vector) vector of column names (from covar) to use for effort covariates for non-iNat PO data}

\item{covs.mean}{(list) list containing one vector for each dataset that says which columns to use as detection covariates that should be averaged (e.g., water temperature)}

\item{covs.sum}{(list) list containing one vector for each dataset that says which columns to use as detection covariates that should be summed (e.g., survey duration)}

\item{offset.area}{(character vector) vector containing column names to use for area offset for each dataset}

\item{DND.maybe}{(numeric) whether to treat maybe detections as detections (1) or non-detections(0); defaults to 1}

\item{keep.conus.grid.id}{(vector) vector of conus.grid.ids to use to fit the model}

\item{min.visits.incl}{(numeric) minimum number (inclusive) of median visits per site to use n-mixture or occupancy models}
}
\value{
(list) list of species data partially formatted for nimble, needs to be input into data_for_nimble before use in nimble
}
\description{
Format species data for nimble
}
