% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_states.R
\docType{data}
\name{stategrid}
\alias{stategrid}
\title{conus.grid.ids by state}
\format{
A data frame with 543,049 rows and 3 variables:
\describe{
\item{conus.grid.id}{integer from 1 to 531296}
\item{name}{two-digit United States state code}
\item{value}{set to 1}
}
}
\usage{
stategrid
}
\description{
conus.grid.ids by state
}
\keyword{datasets}
