% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Title}
\usage{
summarize_samples(
  samples,
  code,
  data,
  constants,
  inits,
  project = 0,
  WAIC = F,
  all.chains = T,
  coarse.grid,
  cutoff = 0,
  block.out,
  gridkey,
  spatkey = NULL,
  local
)
}
\arguments{
\item{local}{}
}
\description{
Title
}
