% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarize chain output from NIMBLE}
\usage{
summarize_samples(
  samples,
  data,
  constants,
  project = 0,
  coarse.grid,
  cutoff = 0,
  block.out,
  gridkey,
  spatkey = NULL,
  effort = F,
  SLURM = F
)
}
\arguments{
\item{samples}{(list) output from NIMBLE}

\item{data}{(list) data created and formatted for NIMBLE by data_for_NIMBLE()$data}

\item{constants}{(list) constants created and formatted for NIMBLE by data_for_NIMBLE()$constants}

\item{project}{(numeric) the number of projections to calculate}

\item{coarse.grid}{(logical) whether a coarse grid was used for the spatial model}

\item{cutoff}{(numeric) where to cutoff chains in addition to the burnin; Default is 0}

\item{block.out}{(character or numeric) which block is excluded}

\item{gridkey}{(data.frame) contains conus.grid.id and associated grid.ids to use for indexing in NIMBLE}

\item{spatkey}{(data.frame) spatial key generated by make_spatkey() and stored in spatRegion}

\item{effort}{(logical) whether to trace effort parameters (T) or not (F)}

\item{SLURM}{(logical) whether a SLURM-based high-performance computing cluster is being used}
}
\description{
Summarize chain output from NIMBLE
}
