% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_for_nimble.R
\name{z_for_nimble}
\alias{z_for_nimble}
\title{Set up z data for nimble}
\usage{
z_for_nimble(covariates)
}
\arguments{
\item{covariates}{}
}
\value{
A list containing ranges as sf objects
}
\description{
Set up z data for nimble
}
